+++ src/Targets/JsOfOCaml/OCamlHelper.fs
@@ -152,9 +152,6 @@ module Type =
 
   let intf  = str "intf"
 
-  let and_ a b = app (str "and_") [a; b]
-  let or_  a b = app (str "or_")  [a; b]
-
   let string_or t  = app (str "or_string") [t]
   let number_or t  = app (str "or_number") [t]
   let boolean_or t = app (str "or_boolean") [t]
@@ -164,23 +161,19 @@ module Type =
   let array_or elemT t = app (str "or_array") [t; elemT]
   let enum_or cases t = app (str "or_enum") [t; cases]
 
-  let union types =
-    let l = List.length types
-    if l < 1 then failwith "union type with only zero or one type"
-    else
-      let rec go i = function
-        | h :: t when i > 8 -> or_ (go (i-1) t) h
-        | xs -> app (tprintf "union%i" i) xs
-      go l types
+  let rec union = function
+    | [] -> failwith "union type with zero elements"
+    | x :: [] -> x
+    | x1 :: x2 :: x3 :: x4 :: x5 :: x6 :: x7 :: x8 :: rest ->
+      app (str "union8") [x1; x2; x3; x4; x5; x6; x7; union (x8 :: rest)]
+    | xs -> app (tprintf "union%i" (List.length xs)) xs
 
-  let intersection types =
-    let l = List.length types
-    if l < 1 then failwith "union type with only zero or one type"
-    else
-      let rec go i = function
-        | h :: t when i > 8 -> and_ (go (i-1) t) h
-        | xs -> app (tprintf "intersection%i" i) xs
-      go l types
+  let rec intersection = function
+    | [] -> failwith "intersection type with zero elements"
+    | x :: [] -> x
+    | x1 :: x2 :: x3 :: x4 :: x5 :: x6 :: x7 :: x8 :: rest ->
+      app (str "intersection8") [x1; x2; x3; x4; x5; x6; x7; intersection (x8 :: rest)]
+    | xs -> app (tprintf "intersection%i" (List.length xs)) xs
 
 [<RequireQualifiedAccess>]
 module Term =